# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация API задач",
        "version": "1"
    },
    "servers": [
        {
            "url": "http://task.ru/"
        }
    ],
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "operationId": "last-task",
                "requestBody": {
                    "description": "Тело запроса фильтров",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetListTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пример успешного ответа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListTask"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "tasks"
                ],
                "operationId": "create-task",
                "requestBody": {
                    "description": "Тело запроса для создания задачи",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пример успешного ответа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{uuid}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "operationId": "one-task",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid задачи",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пример успешного ответа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "404": {
                        "description": "Не найдено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "tasks"
                ],
                "operationId": "update-task",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid задачи",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Тело запроса для обновления задачи",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пример успешного ответа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "403": {
                        "description": "Нет доступа"
                    },
                    "422": {
                        "description": "Ошибки валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "tasks"
                ],
                "operationId": "delete-task",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid задачи",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пример успешного ответа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "403": {
                        "description": "Нет доступа"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Paginate": {
                "title": "Pagination",
                "description": "Модель пагинации",
                "properties": {
                    "links": {
                        "description": "Ссылки для пагинации",
                        "properties": {
                            "first": {
                                "type": "string"
                            },
                            "last": {
                                "type": "string"
                            },
                            "prev": {
                                "type": "string"
                            },
                            "next": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "description": "Мета информация",
                        "properties": {
                            "current_page": {
                                "type": "integer"
                            },
                            "from": {
                                "type": "integer"
                            },
                            "last_page": {
                                "type": "integer"
                            },
                            "path": {
                                "type": "string"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "to": {
                                "type": "integer"
                            },
                            "total": {
                                "type": "integer"
                            },
                            "links": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "url": {
                                            "type": "string"
                                        },
                                        "label": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "CreateTaskRequest": {
                "title": "CreateTaskRequest",
                "description": "Модель request для создания задачи",
                "properties": {
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "GetListTaskRequest": {
                "title": "ListTaskRequest",
                "description": "Модель request для списка задач",
                "properties": {
                    "title": {
                        "description": "Фильтр по названию задачи",
                        "type": "string"
                    },
                    "updated": {
                        "description": "Сортировка по дате обновления",
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ]
                    },
                    "status": {
                        "description": "Фильтр по статусу",
                        "type": "string",
                        "enum": [
                            "create",
                            "await",
                            "processed",
                            "done"
                        ]
                    },
                    "created": {
                        "description": "Сортировка по дате создания",
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ]
                    },
                    "my": {
                        "description": "Вывод только задач авторизированного пользователя",
                        "type": "boolean"
                    },
                    "columns": {
                        "description": "Список выводимих колонок",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateTaskRequest": {
                "title": "UpdateTaskRequest",
                "description": "Модель request для создания задачи",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "create",
                            "await",
                            "processed",
                            "done"
                        ]
                    }
                },
                "type": "object"
            },
            "ListTask": {
                "title": "ListTaskResponse",
                "description": "Модель response для списка задач",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Paginate"
                    },
                    {
                        "properties": {
                            "data": {
                                "description": "Список задач",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Task": {
                "title": "TaskResponse",
                "description": "Модель response для вывода задачи",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "create",
                            "await",
                            "processed",
                            "done"
                        ]
                    },
                    "created_by": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "tasks",
            "description": "API Endpoints of Tasks"
        }
    ]
}